Render drivers are very different accross vendors, on my main pc with an AMD gpu, rounding wasn't even necessary on d3d11 to avoid choppy movement and badly resized textures. On my school laptop with an intel integrated gpu, using ceil fails to avoid both (but it doesn't in my main pc). Using floor instead makes it work, just like what's used in OpenGL I did this just right next to my little brother wondering around with my PC...

+++ src/core/drivers/direct3d11/d3d11draw.cpp
@@ -26,7 +26,7 @@ bool Instance::clear(float color[3])
 void Instance::flush_queueinstances1(sre::Sampler* texture, const sre::RenderInstance1* instances, size_t instance_count, sre::u32 flags, sre::u32 switch_flags)
 {
     UINT inst_num = m_d1buffer.index / sizeof(sre::RenderInstance1);
-    if (m_d1buffer.append(m_dxdevicecontext, instances, sizeof(*instances)*instance_count))
+    if (m_d1buffer.append(m_dxdevicecontext, instances, static_cast<UINT>(sizeof(*instances)*instance_count)))
     {
         switch_flags |= SRE_RENDER_SWITCHTYPE;    
         inst_num = 0;
